+++ tests/deleteNestedProperty.test.js
@@ -55,6 +55,12 @@ describe('Test function setNestedProperty', () => {
         deleteNestedPropertyTest(obj, 'user.list.1');
         expect(obj.user.list).toEqual(['a', 'c']);
     });
+
+    test('Should delete property using array path input', () => {
+        const obj = { user: { profile: { name: 'John Doe', age: 30 } } };
+        deleteNestedPropertyTest(obj, ['user', 'profile', 'name']);
+        expect(obj).toEqual({ user: { profile: { age: 30 } } });
+    });
 })